# services/payments-svc/api/payments/v1/operations_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.7.1
// - protoc             v6.33.0
// source: api/payments/v1/operations.proto

package paymentsv1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationBankOperationsServiceApproveDefund = "/api.payments.v1.BankOperationsService/ApproveDefund"
const OperationBankOperationsServiceBlockHold = "/api.payments.v1.BankOperationsService/BlockHold"
const OperationBankOperationsServiceFundOmnibus = "/api.payments.v1.BankOperationsService/FundOmnibus"
const OperationBankOperationsServiceGetBankPayment = "/api.payments.v1.BankOperationsService/GetBankPayment"
const OperationBankOperationsServiceGetLiquidity = "/api.payments.v1.BankOperationsService/GetLiquidity"
const OperationBankOperationsServiceListBankPayments = "/api.payments.v1.BankOperationsService/ListBankPayments"
const OperationBankOperationsServiceListHolds = "/api.payments.v1.BankOperationsService/ListHolds"
const OperationBankOperationsServiceRejectHold = "/api.payments.v1.BankOperationsService/RejectHold"
const OperationBankOperationsServiceReleaseHold = "/api.payments.v1.BankOperationsService/ReleaseHold"
const OperationBankOperationsServiceRequestDefund = "/api.payments.v1.BankOperationsService/RequestDefund"

type BankOperationsServiceHTTPServer interface {
	// ApproveDefund ApproveDefund is the checker; never the requester, always the same bank.
	ApproveDefund(context.Context, *ApproveDefundRequest) (*Defund, error)
	// BlockHold BlockHold: a sanctioned party has an interest in the funds. They move to
	// the bank's blocked account (BLCK) and a report is due to OFAC within 10
	// business days (31 CFR 501.603).
	BlockHold(context.Context, *HoldDecisionRequest) (*Payment, error)
	// FundOmnibus FundOmnibus sends reserves into the joint account: Fedwire BTRC while
	// Fedwire is open, FedNow LMT1 while it is closed.
	FundOmnibus(context.Context, *FundOmnibusRequest) (*FundOmnibusResponse, error)
	// GetBankPayment GetBankPayment returns one of the bank's clients' payments, with staff detail.
	GetBankPayment(context.Context, *GetBankPaymentRequest) (*Payment, error)
	// GetLiquidity GetLiquidity returns the bank's omnibus position, capacity, alerts,
	// payments waiting for liquidity and defunds.
	GetLiquidity(context.Context, *GetLiquidityRequest) (*Liquidity, error)
	// ListBankPayments ListBankPayments lists the bank's clients' payments, with staff detail.
	ListBankPayments(context.Context, *ListBankPaymentsRequest) (*ListBankPaymentsResponse, error)
	// ListHolds ListHolds lists payments held for sanctions review, and blocked or
	// rejected cases with their OFAC report dates.
	ListHolds(context.Context, *ListHoldsRequest) (*ListHoldsResponse, error)
	// RejectHold RejectHold: prohibited, but no sanctioned party has an interest. Nothing
	// moves (RR04); reported within 10 business days (31 CFR 501.604).
	RejectHold(context.Context, *HoldDecisionRequest) (*Payment, error)
	// ReleaseHold ReleaseHold clears a false positive; the payment proceeds.
	ReleaseHold(context.Context, *HoldDecisionRequest) (*Payment, error)
	// RequestDefund RequestDefund asks for free position back to the master account (maker).
	RequestDefund(context.Context, *RequestDefundRequest) (*Defund, error)
}

func RegisterBankOperationsServiceHTTPServer(s *http.Server, srv BankOperationsServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/v1/ops/payments", _BankOperationsService_ListBankPayments0_HTTP_Handler(srv))
	r.GET("/v1/ops/payments/{id}", _BankOperationsService_GetBankPayment0_HTTP_Handler(srv))
	r.GET("/v1/ops/holds", _BankOperationsService_ListHolds0_HTTP_Handler(srv))
	r.POST("/v1/ops/holds/{id}/release", _BankOperationsService_ReleaseHold0_HTTP_Handler(srv))
	r.POST("/v1/ops/holds/{id}/block", _BankOperationsService_BlockHold0_HTTP_Handler(srv))
	r.POST("/v1/ops/holds/{id}/reject", _BankOperationsService_RejectHold0_HTTP_Handler(srv))
	r.GET("/v1/ops/liquidity", _BankOperationsService_GetLiquidity0_HTTP_Handler(srv))
	r.POST("/v1/ops/funding", _BankOperationsService_FundOmnibus0_HTTP_Handler(srv))
	r.POST("/v1/ops/defunds", _BankOperationsService_RequestDefund0_HTTP_Handler(srv))
	r.POST("/v1/ops/defunds/{id}/approve", _BankOperationsService_ApproveDefund0_HTTP_Handler(srv))
}

func _BankOperationsService_ListBankPayments0_HTTP_Handler(srv BankOperationsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListBankPaymentsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBankOperationsServiceListBankPayments)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListBankPayments(ctx, req.(*ListBankPaymentsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListBankPaymentsResponse)
		return ctx.Result(200, reply)
	}
}

func _BankOperationsService_GetBankPayment0_HTTP_Handler(srv BankOperationsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetBankPaymentRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBankOperationsServiceGetBankPayment)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetBankPayment(ctx, req.(*GetBankPaymentRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Payment)
		return ctx.Result(200, reply)
	}
}

func _BankOperationsService_ListHolds0_HTTP_Handler(srv BankOperationsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListHoldsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBankOperationsServiceListHolds)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListHolds(ctx, req.(*ListHoldsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListHoldsResponse)
		return ctx.Result(200, reply)
	}
}

func _BankOperationsService_ReleaseHold0_HTTP_Handler(srv BankOperationsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in HoldDecisionRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBankOperationsServiceReleaseHold)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ReleaseHold(ctx, req.(*HoldDecisionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Payment)
		return ctx.Result(200, reply)
	}
}

func _BankOperationsService_BlockHold0_HTTP_Handler(srv BankOperationsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in HoldDecisionRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBankOperationsServiceBlockHold)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.BlockHold(ctx, req.(*HoldDecisionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Payment)
		return ctx.Result(200, reply)
	}
}

func _BankOperationsService_RejectHold0_HTTP_Handler(srv BankOperationsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in HoldDecisionRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBankOperationsServiceRejectHold)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RejectHold(ctx, req.(*HoldDecisionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Payment)
		return ctx.Result(200, reply)
	}
}

func _BankOperationsService_GetLiquidity0_HTTP_Handler(srv BankOperationsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetLiquidityRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBankOperationsServiceGetLiquidity)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetLiquidity(ctx, req.(*GetLiquidityRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Liquidity)
		return ctx.Result(200, reply)
	}
}

func _BankOperationsService_FundOmnibus0_HTTP_Handler(srv BankOperationsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in FundOmnibusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBankOperationsServiceFundOmnibus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.FundOmnibus(ctx, req.(*FundOmnibusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*FundOmnibusResponse)
		return ctx.Result(200, reply)
	}
}

func _BankOperationsService_RequestDefund0_HTTP_Handler(srv BankOperationsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in RequestDefundRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBankOperationsServiceRequestDefund)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.RequestDefund(ctx, req.(*RequestDefundRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Defund)
		return ctx.Result(200, reply)
	}
}

func _BankOperationsService_ApproveDefund0_HTTP_Handler(srv BankOperationsServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ApproveDefundRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationBankOperationsServiceApproveDefund)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ApproveDefund(ctx, req.(*ApproveDefundRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*Defund)
		return ctx.Result(200, reply)
	}
}

type BankOperationsServiceHTTPClient interface {
	ApproveDefund(ctx context.Context, req *ApproveDefundRequest, opts ...http.CallOption) (rsp *Defund, err error)
	BlockHold(ctx context.Context, req *HoldDecisionRequest, opts ...http.CallOption) (rsp *Payment, err error)
	FundOmnibus(ctx context.Context, req *FundOmnibusRequest, opts ...http.CallOption) (rsp *FundOmnibusResponse, err error)
	GetBankPayment(ctx context.Context, req *GetBankPaymentRequest, opts ...http.CallOption) (rsp *Payment, err error)
	GetLiquidity(ctx context.Context, req *GetLiquidityRequest, opts ...http.CallOption) (rsp *Liquidity, err error)
	ListBankPayments(ctx context.Context, req *ListBankPaymentsRequest, opts ...http.CallOption) (rsp *ListBankPaymentsResponse, err error)
	ListHolds(ctx context.Context, req *ListHoldsRequest, opts ...http.CallOption) (rsp *ListHoldsResponse, err error)
	RejectHold(ctx context.Context, req *HoldDecisionRequest, opts ...http.CallOption) (rsp *Payment, err error)
	ReleaseHold(ctx context.Context, req *HoldDecisionRequest, opts ...http.CallOption) (rsp *Payment, err error)
	RequestDefund(ctx context.Context, req *RequestDefundRequest, opts ...http.CallOption) (rsp *Defund, err error)
}

type BankOperationsServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewBankOperationsServiceHTTPClient(client *http.Client) BankOperationsServiceHTTPClient {
	return &BankOperationsServiceHTTPClientImpl{client}
}

func (c *BankOperationsServiceHTTPClientImpl) ApproveDefund(ctx context.Context, in *ApproveDefundRequest, opts ...http.CallOption) (*Defund, error) {
	var out Defund
	pattern := "/v1/ops/defunds/{id}/approve"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBankOperationsServiceApproveDefund))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *BankOperationsServiceHTTPClientImpl) BlockHold(ctx context.Context, in *HoldDecisionRequest, opts ...http.CallOption) (*Payment, error) {
	var out Payment
	pattern := "/v1/ops/holds/{id}/block"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBankOperationsServiceBlockHold))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *BankOperationsServiceHTTPClientImpl) FundOmnibus(ctx context.Context, in *FundOmnibusRequest, opts ...http.CallOption) (*FundOmnibusResponse, error) {
	var out FundOmnibusResponse
	pattern := "/v1/ops/funding"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBankOperationsServiceFundOmnibus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *BankOperationsServiceHTTPClientImpl) GetBankPayment(ctx context.Context, in *GetBankPaymentRequest, opts ...http.CallOption) (*Payment, error) {
	var out Payment
	pattern := "/v1/ops/payments/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBankOperationsServiceGetBankPayment))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *BankOperationsServiceHTTPClientImpl) GetLiquidity(ctx context.Context, in *GetLiquidityRequest, opts ...http.CallOption) (*Liquidity, error) {
	var out Liquidity
	pattern := "/v1/ops/liquidity"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBankOperationsServiceGetLiquidity))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *BankOperationsServiceHTTPClientImpl) ListBankPayments(ctx context.Context, in *ListBankPaymentsRequest, opts ...http.CallOption) (*ListBankPaymentsResponse, error) {
	var out ListBankPaymentsResponse
	pattern := "/v1/ops/payments"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBankOperationsServiceListBankPayments))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *BankOperationsServiceHTTPClientImpl) ListHolds(ctx context.Context, in *ListHoldsRequest, opts ...http.CallOption) (*ListHoldsResponse, error) {
	var out ListHoldsResponse
	pattern := "/v1/ops/holds"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationBankOperationsServiceListHolds))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *BankOperationsServiceHTTPClientImpl) RejectHold(ctx context.Context, in *HoldDecisionRequest, opts ...http.CallOption) (*Payment, error) {
	var out Payment
	pattern := "/v1/ops/holds/{id}/reject"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBankOperationsServiceRejectHold))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *BankOperationsServiceHTTPClientImpl) ReleaseHold(ctx context.Context, in *HoldDecisionRequest, opts ...http.CallOption) (*Payment, error) {
	var out Payment
	pattern := "/v1/ops/holds/{id}/release"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBankOperationsServiceReleaseHold))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *BankOperationsServiceHTTPClientImpl) RequestDefund(ctx context.Context, in *RequestDefundRequest, opts ...http.CallOption) (*Defund, error) {
	var out Defund
	pattern := "/v1/ops/defunds"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationBankOperationsServiceRequestDefund))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}